+++ src/features/Percentage.cpp
@@ -57,7 +57,7 @@ class $modify(SliderUI, EditorUI) {
             m_positionSlider->addChild(perLabel);
         }   
         auto thumbPos = m_positionSlider->m_touchLogic->m_thumb->getPosition();
-        perLabel->setString(fmt::format("{:.0f}%", calcPercentage()).c_str());
+        perLabel->setString(fmt::format(Mod::get()->getSettingValue<bool>("use-time-for-percentage") ? "{:.0f} sec" : "{:.0f}%", calcPercentage()).c_str());
         perLabel->setPosition(thumbPos.x, thumbPos.y - 25.f);
         perLabel->setOpacity(255);
         perLabel->stopAllActions();
@@ -66,6 +66,12 @@ class $modify(SliderUI, EditorUI) {
             CCFadeOut::create(.5f),
             nullptr
         ));
+    }
+
+    float timeForXPos(float xPos) {
+        return reinterpret_cast<float(__vectorcall*)(LevelEditorLayer*, float, float)>(
+            geode::base::get() + 0x9c7d0
+            )(LevelEditorLayer::get(), 1907.0f, xPos);
     }
 
     float calcPercentage() {
@@ -78,6 +84,10 @@ class $modify(SliderUI, EditorUI) {
         if (max == 0.f) return 0.f;
         auto winSize = CCDirector::get()->getWinSize();
         auto pos = m_editorLayer->m_objectLayer->convertToNodeSpace(winSize / 2);
+
+        if (Mod::get()->getSettingValue<bool>("use-time-for-percentage"))
+            return LevelTools::timeForXPos(pos.x, CCArray::create(), 1);
+
         return pos.x / max * 100;
     }
 };